refactor: split tabs component, fix opened file persistance
